fix: make the build not die if window.fetch isn't available.
