+++ package.json
@@ -100,7 +100,7 @@
     "mime": "^2.4.2",
     "multer": "^1.4.1",
     "mustache": "^4.1.0",
-    "node-fetch": "^2.6.7",
+    "node-fetch": "^3.2.1",
     "nodemailer": "^6.5.0",
     "owasp-password-strength-test": "^1.3.0",
     "parse-database-url": "^0.3.0",
